+++ src/scicloj/metamorph/ml.clj
@@ -16,6 +16,7 @@
    [tech.v3.datatype.errors :as errors]
    [tech.v3.datatype.export-symbols :as exporter]
    [tech.v3.datatype.functional :as dfn]
+   [scicloj.metamorph.ml.tidy-models :as tidy]
    [clojure.set :as set])
     ;;
 
@@ -663,18 +664,6 @@
       ;;   (vec (distinct simple-predicted-values))
       ;;   (-> target-cat-maps-from-predict vals first :lookup-table)))
 
-(defn- assoc-categorical-maps [pred-ds target-categorical-map target-columns]
-  (if target-categorical-map
-    (reduce (fn [ds col]
-              (ds/assoc-metadata
-               ds
-               [col]
-               :categorical-map (get target-categorical-map col)))
-            pred-ds
-            target-columns
-            )
-    pred-ds))
-
 
 (defn predict
   "Predict returns a dataset with only the predictions in it.
@@ -694,17 +683,13 @@
   (let [{:keys [predict-fn] :as model-def} (options->model-def (:options model))
         feature-ds (ds/select-columns dataset (:feature-columns model))
         thawed-model (thaw-model model model-def)
-        pred-ds 
-        (-> 
-         (predict-fn feature-ds
-                     thawed-model
-                     model)
-         (assoc-categorical-maps 
-                                 (:target-categorical-maps model)
-                                 (:target-columns model)))]
+        pred-ds (predict-fn feature-ds
+                            thawed-model
+                            model)]
 
     (warn-inconsitent-maps model pred-ds)
     pred-ds))
+
 
 (defn loglik [model y yhat]
 
@@ -731,9 +716,13 @@
         (get
          (options->model-def (:options model))
          :tidy-fn)]
+
     (if tidy-fn
-      (tidy-fn model)
+      (tidy/validate-tidy-ds
+       (tidy-fn model))
       (ds/->dataset {}))))
+
+
 
 (defn glance
   "Gives a glance on the model, returning a dataset with model information
@@ -754,8 +743,10 @@
          (options->model-def (:options model))
          :glance-fn)]
     (if glance-fn
-      (glance-fn model)
+      (tidy/validate-glance-ds
+       (glance-fn model))
       (ds/->dataset {}))))
+
 
 
 (defn augment
@@ -779,7 +770,9 @@
          (options->model-def (:options model))
          :augment-fn)]
     (if augment-fn
-      (augment-fn model data)
+      (tidy/validate-augment-ds
+       (augment-fn model data)
+       data)
       data)))