+++ build.clj
@@ -8,10 +8,10 @@
    [clojure.tools.build.api :as b] ; for b/git-count-revs
    [org.corfield.build :as bb]))
 
-(def lib 'scicloj/metamorph.ml)
+(def lib 'org.scicloj/metamorph.ml)
 ; alternatively, use MAJOR.MINOR.COMMITS:
 ;; (def version (format "6.2.%s" (b/git-count-revs nil)))
-(def version "0.8.2")
+(def version "0.9.0")
 (def class-dir "target/classes")
 (def basis (b/create-basis {:project "deps.edn"}))
 (def jar-file (format "target/%s-%s.jar" (name lib) version))