+++ CHANGELOG.md
@@ -2,6 +2,7 @@ unreleased
 
 - added fastmath/lm
 - added tidy output validation
+- fixed #19
 
 0.8.2
  -fixed metric bug